+++ deepinv/loss/metric/__init__.py
@@ -13,5 +13,5 @@ from .distortion import (
     ERGAS,
     HaarPSI,
 )
-from .perceptual import NIQE, LPIPS
+from .perceptual import NIQE, LPIPS, BlurStrength, SharpnessIndex
 from .functional import cal_mse, cal_psnr, cal_mae, signal_noise_ratio