+++ docs/source/conf.py
@@ -345,6 +345,7 @@ sphinx_gallery_conf = {
             "../../examples/plug-and-play",
             "../../examples/sampling",
             "../../examples/unfolded",
+            "../../examples/blind-inverse-problems",
             "../../examples/self-supervised-learning",
             "../../examples/adversarial-learning",
             "../../examples/external-libraries",